Rename `mod common` -> `stdio`
