+++ src/lib.rs
@@ -23,11 +23,11 @@ pub mod blackhole;
 pub(crate) mod block;
 pub mod captures;
 pub(crate) mod codec;
-mod common;
 pub mod config;
 pub mod generator;
 pub mod inspector;
 pub mod observer;
 pub(crate) mod payload;
 pub mod signals;
+mod stdio;
 pub mod target;
